fj_math

Struct Point

#[repr(C)]
pub struct Point<const D: usize> {
    pub coords: Vector<D>,
}

An n-dimensional point

The dimensionality of the point is defined by the const generic D parameter.

Fields
coords: Vector<D>

The coordinates of the point

Implementations

impl<const D: usize> Point<D>

pub fn origin() -> Self

Construct a Point at the origin of the coordinate system

pub fn from_array(array: [f64; D]) -> Self

Construct a Point from an array

pub fn from_na(point: Point<f64, D>) -> Self

Construct a Point from an nalgebra vector

pub fn to_na(self) -> Point<f64, D>

Convert the point into an nalgebra point

pub fn to_xyz(self) -> Point<3>

Convert the point to a 3-dimensional point

See Vector::to_xyz for details. This method follows the same rules.

pub fn distance_to(&self, other: &Self) -> Scalar

Compute the distance between two points

pub fn floor(self) -> Self

Transform the point by calling Scalar::floor on its components