fj_math

Struct Scalar

pub struct Scalar { /* private fields */ }

A rational, finite scalar value

Scalar is a wrapper around f64 which guarantees that the contained value is both rational and finite. This allows Scalar to provide implementations of Eq, Ord, and Hash; enabling Scalar (and types built on top of it), to be used as keys in various sets and maps.

Implementations

impl Scalar

pub const ZERO: Self

The Scalar instance that represents zero

pub const ONE: Self

The Scalar instance that represents one

pub const TWO: Self

The Scalar instance that represents two

pub const MIN: Self

The smallest Scalar value

pub const MAX: Self

The largest Scalar value

pub const PI: Self

The Scalar instance that represents π (pi)

pub const TAU: Self

The Scalar instance that represents τ (tau)

pub fn from_f64(value: f64) -> Self

Construct a Scalar from an f64

Panics

Panics, if the value provided is NaN or infinite.

pub fn into_f32(self) -> f32

Convert the scalar value into an f32

pub fn into_f64(self) -> f64

Convert the scalar value into an f64

pub fn into_u64(self) -> u64

Convert the scalar value into a u64

pub fn is_negative(self) -> bool

Indicate whether the scalar value is negative

pub fn is_positive(self) -> bool

Indicate whether the scalar value is positive

pub fn is_zero(self) -> bool

Indicate whether the scalar value is zero

pub fn sign(self) -> Sign

Compute the sign of the scalar

pub fn abs(self) -> Self

Compute the absolute scalar value

pub fn max(self, other: impl Into<Self>) -> Self

Compute the maximum of this and another scalar value

pub fn floor(self) -> Self

Compute the largest integer smaller than or equal to the scalar value

pub fn ceil(self) -> Self

Compute the smallest integer larger than or equal to the scalar value

pub fn round(self) -> Self

Round the scalar value

pub fn sqrt(self) -> Self

Compute the square root of the scalar value

pub fn sin(self) -> Self

Compute the sine of the scalar value

pub fn cos(self) -> Self

Compute the cosine of the scalar value

pub fn sin_cos(self) -> (Self, Self)

Compute sine and cosine of the scalar value

pub fn acos(self) -> Self

Compute the arccosine of the scalar value

pub fn atan2(self, other: Self) -> Self

Compute the four-quadrant arctangent of the scalar value