fj_math

Struct Vector

#[repr(C)]
pub struct Vector<const D: usize> {
    pub components: [Scalar; D],
}

An n-dimensional vector

The dimensionality of the vector is defined by the const generic D parameter.

Fields
components: [Scalar; D]

The components of the vector

Implementations

impl<const D: usize> Vector<D>

pub fn from_component(scalar: impl Into<Scalar>) -> Self

Construct a vector whose components are all equal

pub fn to_na(self) -> SVector<f64, D>

Convert the vector into an nalgebra vector

pub fn to_t(self) -> Vector<1>

Convert the vector to a 1-dimensional vector

pub fn to_uv(self) -> Vector<2>

Convert the vector into a 2-dimensional vector

If the vector is 0-, or 1-dimensional, the missing components will be initialized to zero.

If the vector has higher dimensionality than two, the superfluous components will be discarded.

pub fn to_xyz(self) -> Vector<3>

Convert the vector into a 3-dimensional vector

If the vector is 0-, 1-, or 2-dimensional, the missing components will be initialized to zero.

If the vector has higher dimensionality than three, the superfluous components will be discarded.

pub fn magnitude(&self) -> Scalar

Compute the magnitude of the vector

pub fn normalize(&self) -> Self

Compute a normalized version of the vector

pub fn angle_to(&self, other: &Self) -> Scalar

Compute the angle between this vector and another

Returns a zero angle, if the magnitude of self or other is zero.

pub fn dot(&self, other: &Self) -> Scalar

Compute the dot product with another vector

pub fn outer(&self, other: &Self) -> Bivector<D>

Compute the outer product with another vector

pub fn scalar_projection_onto(&self, other: &Self) -> Scalar

Compute the scalar projection of this vector onto another

pub fn floor(self) -> Self

Transform the vector by calling Scalar::floor on its components

impl Vector<1>

pub fn unit_t() -> Self

Construct a Vector that represents the t-axis

impl Vector<2>

pub fn unit_u() -> Self

Construct a Vector that represents the u-axis

pub fn unit_v() -> Self

Construct a Vector that represents the v-axis

pub fn cross2d(&self, other: &Self) -> Scalar

Compute the 2D cross product with another vector

pub fn is_between(&self, others: [impl Into<Self>; 2]) -> bool

Determine whether this vector is between two other vectors

pub fn to_perpendicular(&self) -> Self

Compute a vector that is perpendicular to this one

impl Vector<3>

pub fn unit_x() -> Self

Construct a Vector that represents the x-axis

pub fn unit_y() -> Self

Construct a Vector that represents the y-axis

pub fn unit_z() -> Self

Construct a Vector that represents the z-axis

pub fn cross(&self, other: &Self) -> Self

Compute the cross product with another vector

pub fn xy(&self) -> Vector<2>

Construct a new vector from this vector’s x and y components